+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- Make `FileStorage` to check staleness of all cache files with set interval. (#169)
+- Support AWS S3 storages. (#164)
+- Move `typing_extensions` from requirements.txt to pyproject.toml. (#161)
+
 ## 0.0.21 (29th December, 2023)
 
 - Fix inner transport and connection pool instances closing. (#147)